% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/szlaczki.R
\name{szlaczki}
\alias{szlaczki}
\title{Robi szlaczki}
\usage{
szlaczki(
  tekst,
  wys = 7,
  szer = 79,
  symbole = szlaczki:::marks(),
  linie = 2,
  druk = FALSE,
  ...
)
}
\arguments{
\item{tekst}{String, Tekst do umieszczenia w szlaczku - znajdzie sie na środku po lewo.}

\item{wys}{Liczba, Wysokosc szlaczka - domyslnie 7.}

\item{szer}{Liczba, Szerokosc szlaczka - domyslnie 79.}

\item{symbole}{Wektor symboli do napisania szlaczka - domyslnie zbior wszystkich symboli ASCII. Jezeli zostanie podana liczba n, wybranych zostanie losowo n symboli sposrod wszystkich symboli ASCII.}

\item{linie}{Liczba, Co ile linii ma sie pojawiac linia wypelniona symbolami. Jezeli zostanie podana linie = 1 lub wartosc logiczna FALSE, wszystkie linie beda wypelnione.}

\item{druk}{Logiczna, Czy wydrukowac szlaczek do konsoli.}
}
\value{
Szlaczek skopiowany do schowka (Windows) badz wydrukowany do konsoli.
}
\description{
Robi szlaczki
}
